WIP: Criada validação de position.name; Teste de mensagens de erro
