+++ db/migrate/20230615144359_add_index_to_positions_code.rb
@@ -1,0 +1,5 @@
+class AddIndexToPositionsCode < ActiveRecord::Migration[7.0]
+  def change
+    add_index :positions, :code, unique: true
+  end
+end
